+++ src/scanoss/cyclonedx.py
@@ -28,6 +28,9 @@ import os.path
 import sys
 import uuid
 
+from cyclonedx.schema import SchemaVersion
+from cyclonedx.validation.json import JsonValidator
+
 from . import __version__
 from .scanossbase import ScanossBase
 from .spdxlite import SpdxLite
@@ -386,7 +389,45 @@ class CycloneDx(ScanossBase):
             'unknown': 'unknown',
         }.get(value, 'unknown')
 
+    def is_cyclonedx_json(self, json_string: str) -> bool:
+        """
+        Validate if the given JSON string is a valid CycloneDX JSON string
 
+        Args:
+            json_string (str): JSON string to validate
+        Returns:
+            bool: True if the JSON string is valid, False otherwise
+        """
+        try:
+            cdx_json_validator = JsonValidator(SchemaVersion.V1_6)
+            json_validation_errors = cdx_json_validator.validate_str(json_string)
+            if json_validation_errors:
+                self.print_stderr(f'ERROR: Problem parsing input JSON: {json_validation_errors}')
+                return False
+            return True
+        except Exception as e:
+            self.print_stderr(f'ERROR: Problem parsing input JSON: {e}')
+            return False
+
+    def get_purls_request_from_cdx(self, cdx_dict: dict) -> dict:
+        """
+        Get the list of PURL requests (purl + requirement) from the given CDX dictionary
+
+        Args:
+            cdx_dict (dict): CDX dictionary to parse
+        Returns:
+            list[dict]: List of PURL requests (purl + requirement)
+        """
+        components = cdx_dict.get('components', [])
+        parsed_purls = []
+        for component in components:
+            version = component.get('version')
+            if version:
+                parsed_purls.append({'purl': component.get('purl'), 'requirement': version})
+            else:
+                parsed_purls.append({'purl': component.get('purl')})
+        purl_request = {'purls': parsed_purls}
+        return purl_request
 #
 # End of CycloneDX Class
 #